+++ js/_autoload_patch.js
@@ -1,0 +1,2 @@
+<script src="/js/indexeddb.js"></script>
+<script src="/js/reports.js"></script>